extract RankingEntry interface out of daily scoreboard component
