+++ src/@types/RankingEntry.ts
@@ -1,0 +1,6 @@
+export interface RankingEntry {
+  rank: number;
+  username: string;
+  score: number;
+  wave: number;
+}
