+++ src/plugins/api/daily-api.ts
@@ -1,6 +1,6 @@
 import type { GetDailyRankingsPageCountRequest, GetDailyRankingsRequest } from "#app/@types/DailyApi";
 import { ApiBase } from "#app/plugins/api/api-base";
-import type { RankingEntry } from "#app/ui/components/daily-run-scoreboard";
+import type { RankingEntry } from "#app/@types/RankingEntry";
 
 /**
  * A wrapper for daily-run API requests.
