+++ test/plugins/api/daily-api.test.ts
@@ -1,7 +1,7 @@
 import type { GetDailyRankingsPageCountRequest, GetDailyRankingsRequest } from "#app/@types/DailyApi";
 import { DailyApi } from "#app/plugins/api/daily-api";
 import { getApiBaseUrl } from "#test/test-utils/testUtils";
-import { type RankingEntry } from "#app/ui/components/daily-run-scoreboard";
+import type { RankingEntry } from "#app/@types/RankingEntry";
 import { ScoreboardCategory } from "#enums/scoreboard-category";
 import { http, HttpResponse } from "msw";
 import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
